Rename iOS arguments to Apple arguments
