+++ src/Microsoft.DotNet.XHarness.CLI/Commands/iOS/AppleTestCommand.cs
@@ -7,7 +7,7 @@ using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using Microsoft.DotNet.XHarness.Apple;
-using Microsoft.DotNet.XHarness.CLI.CommandArguments.iOS;
+using Microsoft.DotNet.XHarness.CLI.CommandArguments.Apple;
 using Microsoft.DotNet.XHarness.Common.CLI;
 using Microsoft.DotNet.XHarness.Common.Logging;
 using Microsoft.DotNet.XHarness.iOS;
@@ -27,11 +27,11 @@ namespace Microsoft.DotNet.XHarness.CLI.Commands.iOS
     {
         private const string CommandHelp = "Runs a given iOS/tvOS/watchOS test application bundle containing TestRunner in a target device/simulator";
 
-        private readonly iOSTestCommandArguments _arguments = new iOSTestCommandArguments();
+        private readonly AppleTestCommandArguments _arguments = new AppleTestCommandArguments();
 
         protected override string CommandUsage { get; } = "ios test [OPTIONS] [-- [RUNTIME ARGUMENTS]]";
         protected override string CommandDescription { get; } = CommandHelp;
-        protected override iOSAppRunArguments iOSRunArguments => _arguments;
+        protected override AppleAppRunArguments iOSRunArguments => _arguments;
 
         public AppleTestCommand() : base("test", false, CommandHelp)
         {
